refactor: edit app architecture
