+++ src/commands/fs/cp.js
@@ -1,8 +1,8 @@
-import fs from 'fs';
-import { cwd } from 'process';
-import { resolve } from 'node:path';
-import { displayErrorMessage } from '../utils/messages.js';
-import { log } from '../utils/coloredMsg.js';
+import fs from "fs";
+import { cwd } from "process";
+import { resolve } from "node:path";
+import { displayErrorMessage } from "../../utils/messages.js";
+import { log } from "../../utils/coloredMsg.js";
 
 export const cp = async (srcPath, destPath) => {
   const sourceFilePath = resolve(cwd(), srcPath);
@@ -17,17 +17,17 @@ export const cp = async (srcPath, destPath) => {
     const readStream = fs.createReadStream(sourceFilePath);
     const writeStream = fs.createWriteStream(destinationFilePath);
 
-    readStream.on('error', (error) => {
+    readStream.on("error", (error) => {
       displayErrorMessage(error.message);
     });
 
-    writeStream.on('error', (error) => {
+    writeStream.on("error", (error) => {
       displayErrorMessage(error.message);
     });
 
     readStream.pipe(writeStream);
 
-    readStream.on('end', () => {
+    readStream.on("end", () => {
       log.green(`File ${srcPath} copied to ${destPath} successfully`);
     });
   });
