+++ src/commands/fs/rm.js
@@ -1,8 +1,8 @@
-import fs from 'fs';
-import { resolve } from 'node:path';
-import { cwd } from 'process';
-import { displayErrorMessage } from '../utils/messages.js';
-import { log } from '../utils/coloredMsg.js';
+import fs from "fs";
+import { resolve } from "node:path";
+import { cwd } from "process";
+import { displayErrorMessage } from "../../utils/messages.js";
+import { log } from "../../utils/coloredMsg.js";
 
 export const rm = async (filePath) => {
   const fileToDelete = resolve(cwd(), filePath);
