+++ src/commands/index.js
@@ -1,14 +1,13 @@
-export { add } from "./add.js";
-export { cat } from "./cat.js";
-export { cd } from "./cd.js";
-export { compress } from "./compress.js";
-export { cp } from "./cp.js";
-export { decompress } from "./decompress.js";
-export { hash } from "./hash.js";
-export { ls } from "./ls.js";
-export { mv } from "./mv.js";
-export { printOsInfo } from "./os.js";
-export { rm } from "./rm.js";
-export { rn } from "./rn.js";
-export { up } from "./up.js";
-
+export { add } from "./fs/add.js";
+export { cat } from "./fs/cat.js";
+export { cd } from "./nwd/cd.js";
+export { compress } from "./compress/compress.js";
+export { cp } from "./fs/cp.js";
+export { decompress } from "./compress/decompress.js";
+export { hash } from "./hash/hash.js";
+export { ls } from "./nwd/ls.js";
+export { mv } from "./fs/mv.js";
+export { printOsInfo } from "./os/os.js";
+export { rm } from "./fs/rm.js";
+export { rn } from "./fs/rn.js";
+export { up } from "./nwd/up.js";
