+++ src/commands/nwd/cd.js
@@ -1,7 +1,7 @@
-import { resolve } from 'node:path';
-import { chdir } from 'node:process';
-import { cwd } from 'process';
-import { displayErrorMessage } from '../utils/messages.js';
+import { resolve } from "node:path";
+import { chdir } from "node:process";
+import { cwd } from "process";
+import { displayErrorMessage } from "../../utils/messages.js";
 
 export const cd = async (path) => {
   const currentDir = cwd();
