+++ src/commands/nwd/ls.js
